refactor :: auth code generator
