+++ src/main/java/com/daemawiki/daemawiki/domain/mail/service/UserMailSendService.java
@@ -7,7 +7,7 @@ import com.daemawiki.daemawiki.domain.mail.event.model.MailSendEvent;
 import com.daemawiki.daemawiki.domain.mail.event.handler.UserMailSendEventHandler;
 import com.daemawiki.daemawiki.domain.mail.usecase.UserMailSendUseCase;
 import com.daemawiki.daemawiki.domain.user.repository.UserRepository;
-import com.daemawiki.daemawiki.global.utils.AuthCodeGenerator;
+import com.daemawiki.daemawiki.global.utils.crypto.AuthCodeGenerator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
