+++ src/main/java/com/daemawiki/daemawiki/global/utils/crypto/AuthCodeGenerator.java
@@ -1,4 +1,4 @@
-package com.daemawiki.daemawiki.global.utils;
+package com.daemawiki.daemawiki.global.utils.crypto;
 
 import org.springframework.stereotype.Component;
